requirements.txt
django
djangorestframework
pillow
django-cors-headers
opencv-python-headless
requests
numpy
daphne
django-channels
face_recognition